move : MainActivity와 함수들 분리
작업 완료
LoGolGi.kt -> 메인 앱 함수
LoadingView.kt -> 로딩화면
ScreenOrientation.kt -> 화면 가로고정 함수

작업 예정
메인화면
초기화시 숫자가 안바뀌는 현상

LazyColumn 목록과 HorizontalPager로 변경 가능하게 하기

골드세팅
뷰모델 해결해보기

캐릭터 상세
viewModel에서 사용되는 공통 함수 model로 빼기

장비
빈값에 빈값인 이미지보여주기
-> list에서 find로 찾고 없다면 빈값, 있다면 채우기 이런식?
(나중에) 팔찌 상중하 옵션별로 색부여

코드 전체
하드 코딩된 String값 정리
공통 함수 정리 및 viewModel 정리
