+++ app/src/main/java/com/hongmyeoun/goldcalc/view/logolgiMain/ScreenOrientation.kt
@@ -1,0 +1,13 @@
+package com.hongmyeoun.goldcalc.view.logolgiMain
+
+import android.app.Activity
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+
+// 화면 세로 고정
+@Composable
+fun LockScreenOrientation(orientation: Int) {
+    val context = LocalContext.current
+    val activity = context as? Activity
+    activity?.requestedOrientation = orientation
+}
